Modification of
L2B
L2BHeader
WindSwath
WVC

objects to allow us to attach an unsigned 4 byte int as a quality flag and
a 4 byte float as a rain impact parameter in the L2B data for each WVC.

L2BHeader::Read attempts to find a 80 byte char string at top of L2B file,
if found, it parses version id in string, if not sets version id to 1.0.

If version id == 1.0, then it does not try to read the qual flag and
rain impact.  If version id is 2.0 it tries to read the qual flag and
rain impact from L2B data file.

Version ID in L2BHeader::Write is hard-coded to 2.0, meaning that all
new l2b files created will be version 2.0 and have the new paramters.

AGF 6/8/2010